I refactored more the the code for the image manager. I moved all of the code that loads each image into the image store (e.g., ImageManager.loadImage(image_filename)) from the main game class Asteroids and moved it to a method in the ImageManager in order to clean up and reduce the amount of code in the main game class.
I refactored the code responsible for generating UFOs by moving it from the main game class and placing it in the ufo manager. I also modified the code to  use a Factory appraoch.

I also cleaned up some erroneous code in each of the UFO classes. I realized I was doing similar logic in the overridden updatePosition methods in UFOShorty and UFOStringEntity that I was doing in UFOEntity. So I eliminated the overriden methods in UFOShorty and UFOStrongEntity.

I also started to move the code for displaying the HUD in its own method in the Asteroids class.

+++ src/galacticwarreboot/Constants.java
@@ -312,7 +312,7 @@ public class Constants
   public static final long   UFO_TIME_BETWEEN_SHIELD_HITS                     = 750;
 
   public static final long   UFO_SHORTY_TIME_BETWEEN_HITS                     = 50;
-  public static final long   UFO_SHORTY_HIT_POINTS                            = 10;
+  public static final int   UFO_SHORTY_HIT_POINTS                            = 10;
 
   public static final int    SHIP_STARTING_FIREPOWER                          = 2;
   public static final int    SHIP_MAX_FIREPOWER                               = 5;
